backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
schedule==1.2.0
pytz==2023.3
redis[hiredis]==5.0.1
PyJWT==2.8.0
pandas==2.3.2
openpyxl==3.1.5
xlrd==2.0.2
